Fix errors

+++ BESA/AgenteResPwA/ResPwAEntities/src/ResPwAEntities/Controllers/ReglaJpaController.java
@@ -99,7 +99,7 @@ public class ReglaJpaController implements Serializable {
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
             if (msg == null || msg.length() == 0) {
-                BigDecimal id = regla.getId();
+                Long id = regla.getId();
                 if (findRegla(id) == null) {
                     throw new NonexistentEntityException("The regla with id " + id + " no longer exists.");
                 }
@@ -162,7 +162,7 @@ public class ReglaJpaController implements Serializable {
         }
     }
 
-    public Regla findRegla(BigDecimal id) {
+    public Regla findRegla(Long id) {
         EntityManager em = getEntityManager();
         try {
             return em.find(Regla.class, id);
